+++ frontend/src/Pages/Profile/Profile.scss
@@ -14,6 +14,7 @@
   height: fit-content;
   padding: 40px;
   border-radius: 16px;
+  margin-bottom: 309px;
 }
 
 .profile-settings__title {
