+++ frontend/src/utils/api.js
@@ -76,3 +76,62 @@ export const logoutUserAPI = async () => {
     throw new Error(`Logout failed: ${error.message}`);
   }
 };
+
+export const getUserAPI = async () => {
+  const url = `${baseUrl}/users/me/`;
+  const options = {
+    method: 'GET',
+    headers: {
+      ...defaultHeaders,
+      authorization: `Token ${getCookie('token')}`,
+    },
+  };
+  const data = await request(url, options);
+  return data;
+};
+
+export const updateUserAPI = async (userData) => {
+  const url = `${baseUrl}/users/me/`;
+  const options = {
+    method: 'PATCH',
+    headers: {
+      ...defaultHeaders,
+      authorization: `Token ${getCookie('token')}`,
+    },
+    body: JSON.stringify(userData),
+  };
+
+  const data = await request(url, options);
+  return data;
+};
+
+export const deleteUserAPI = async () => {
+  const url = `${baseUrl}/users/me/`;
+  const options = {
+    method: 'DELETE',
+    headers: {
+      ...defaultHeaders,
+      authorization: `Token ${getCookie('token')}`,
+    },
+  };
+
+  try {
+    await request(url, options);
+    deleteCookie('token');
+  } catch (error) {
+    throw new Error(`Delete user failed: ${error.message}`);
+  }
+};
+
+export const getAccontsAPI = async () => {
+  const url = `${baseUrl}/finance/`;
+  const options = {
+    method: 'GET',
+    headers: {
+      ...defaultHeaders,
+      authorization: `Token ${getCookie('token')}`,
+    },
+  };
+  const data = await request(url, options);
+  return data;
+};
